requirements.txt
fastapi
uvicorn
scikit-learn
joblib
numpy
requests
python-dotenv
groq